move app and booklist inside component
