+++ src/components/BooksList/BooksList.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import React, { useEffect, useState } from 'react';
 import { useSelector } from 'react-redux';
-import Book from '../../components/book/index';
+import Book from '../book/index';
 
 const BooksList = () => {
   const [Books, setBooks] = useState([]);
